Move example format from cfg(test) into main code
so that we can use it e.g. from the Web example page.
